requirements.txt
Django>=4.1,<4.2
wagtail>=5.0,<5.1

# Django Oscar
django-oscar>=3.2,<3.3
git+https://github.com/mfatgg/django-oscar-promotions@upgrade#egg=django-oscar-promotions
django-debug-toolbar>=2.2,<4.3
django-extensions>=2.2,<3.3
sorl-thumbnail>=12.9,<12.10
django-environ>=0.4,<0.11
pysolr>=3.9,<3.10
requests>=2.25,<3
pylint>=2.17.4
pylint-django>=2.5.3
black>=23.3.0
pycountry
